#25 Add check for duplicate serial number
Update issuing serial logical to follow that of CRL serial in settings.
Need to update the Certificate Authority to validate this value on load
of CA>

+++ bundles/net.sourceforge.dkartaschew.halimede.core/src/main/java/net/sourceforge/dkartaschew/halimede/data/CertificateAuthority.java
@@ -1162,7 +1162,7 @@ public class CertificateAuthority {
 	 * @throws IOException If storing the CA state fails.
 	 */
 	public synchronized BigInteger getNextSerialNumber() throws IOException {
-		BigInteger bint = BigInteger.valueOf(settings.getAndIncrementSerial());
+		BigInteger bint = settings.getAndIncrementSerial();
 		saveSettings();
 		return bint;
 	}
